Divide javascript code into modules
