+++ src/modules/display.js
@@ -1,42 +1,4 @@
-const likeShow = (array) => {
-  const myArray = array.filter((object) => object.item_id !== '');
-  console.log(myArray);
-  myArray.forEach((object) => {
-    const likes = document.querySelector(`#count${object.item_id}`);
-    if (object.likes === null) {
-      likes.innerHTML = '0 Likes';
-    } else {
-      likes.innerHTML = `${object.likes} Likes`;
-    }
-  });
-};
-
-const getLike = async () => {
-  await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TUbnIqBO6HOTZf6v0kyK/likes/', {
-    method: 'GET',
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  })
-    .then((response) => response.json())
-    .then((json) => likeShow(json));
-};
-
-window.addEventListener('load', () => {
-  getLike();
-});
-
-const likeSend = async (id) => {
-  await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TUbnIqBO6HOTZf6v0kyK/likes/', {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: id,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-};
+import enableButton from './enablebuttons.js';
 
 const display = (array) => {
   const container = document.querySelector('#maincontainer');
@@ -57,14 +19,7 @@ const display = (array) => {
     </div>
     </section>`;
   });
-  const likebuttons = document.querySelectorAll('.likebutton');
-  likebuttons.forEach((button) => {
-    button.addEventListener('click', async (e) => {
-      const identity = e.target.parentNode.id;
-      await likeSend(identity);
-      getLike();
-    });
-  });
+  enableButton();
 };
 
 export default display;
