+++ src/modules/getlike.js
@@ -1,0 +1,14 @@
+import likeShow from './likeshow.js';
+
+const getLike = async () => {
+  await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/TUbnIqBO6HOTZf6v0kyK/likes/', {
+    method: 'GET',
+    headers: {
+      'Content-type': 'application/json; charset=UTF-8',
+    },
+  })
+    .then((response) => response.json())
+    .then((json) => likeShow(json));
+};
+
+export default getLike;
